+++ changelog.md
@@ -8,6 +8,7 @@
 - The default prefix of hostname and tasks are the FrameworkName.
 - Optimize framework suppress
 - Add constraint "hostname eq" to accept only offers with the given hostname.
+- Check if the random generated port is already in use at the given agent.
 
 ## v0.2.0